services/tg_bot/requirements.txt
pyTelegramBotAPI >= 4.5.1
python-dateutil >= 2.8.2
peewee >= 3.14.10
gunicorn >= 20.1.0
flask >= 2.1.2
pymysql
requests
